# openapi.yaml
openapi: 3.0.3
info:
  title: Nexus Teammate API
  version: 0.1.0
  description: |
    MVP OpenAPI draft for Nexus Teammate.
    Tech stack: React + Midwayjs + MySQL.
servers:
  - url: /api/v1
...
paths:
...
  /boards/{boardId}/tasks:
    get:
      tags: [Tasks]
      summary: Get board tasks
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BoardId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Keyword'
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: assigneeId
          in: query
          schema:
            type: integer
            format: int64
        - name: aiState
          in: query
          schema:
            type: string
        - name: takeoverState
          in: query
          schema:
            type: string
        - name: viewMode
          in: query
          schema:
            type: string
            enum: [kanban, list]
            default: kanban
      responses:
        '200':
          description: Board tasks
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BoardTasksResponse'
  /tasks:
    post:
      tags: [Tasks]
      summary: Create task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '200':
          description: Task created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CreateTaskResponse'
  /tasks/{taskId}:
    get:
      tags: [Tasks]
      summary: Get task detail
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TaskDetail'
    put:
      tags: [Tasks]
      summary: Update task
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /tasks/{taskId}/transition:
    post:
      tags: [Tasks]
      summary: Transition task status
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransitionTaskRequest'
      responses:
        '200':
          description: Transition accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /tasks/{taskId}/comments:
    post:
      tags: [Tasks]
      summary: Add task comment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskCommentRequest'
      responses:
        '200':
          description: Comment added
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TaskComment'
  /tasks/{taskId}/takeover:
    post:
      tags: [Tasks]
      summary: Take over or continue task
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskTakeoverRequest'
      responses:
        '200':
          description: Takeover action applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Keyword:
      name: keyword
      in: query
      schema:
        type: string
...
    BoardId:
      name: boardId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
          example: OK
        message:
          type: string
          example: success
        data:
          nullable: true
        requestId:
          type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...
    UserSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    Label:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
    LatestExecutionSummary:
      type: object
      properties:
        executionId:
          type: integer
          format: int64
        agentName:
          type: string
        status:
          type: string
        summary:
          type: string
    TaskCard:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        priority:
          type: string
        aiState:
          type: string
        takeoverState:
          type: string
        assignee:
          $ref: '#/components/schemas/UserSummary'
        latestExecution:
          $ref: '#/components/schemas/LatestExecutionSummary'
    BoardTaskColumn:
      type: object
      properties:
        statusCode:
          type: string
        name:
          type: string
        count:
          type: integer
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskCard'
    BoardTasksResponse:
      type: object
      properties:
        viewMode:
          type: string
        columns:
          type: array
          items:
            $ref: '#/components/schemas/BoardTaskColumn'
        list:
          type: array
          items:
            $ref: '#/components/schemas/TaskCard'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreateTaskRequest:
      type: object
      required: [projectId, boardId, title, sourceType]
      properties:
        projectId:
          type: integer
          format: int64
        boardId:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        priority:
          type: string
        assigneeId:
          type: integer
          format: int64
        sourceType:
          type: string
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        customFields:
          type: object
          additionalProperties: true
    CreateTaskResponse:
      type: object
      properties:
        taskId:
          type: integer
          format: int64
        statusCode:
          type: string
        createdExecutionIds:
          type: array
          items:
            type: integer
            format: int64
    TaskComment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        taskId:
          type: integer
          format: int64
        authorType:
          type: string
        authorName:
          type: string
        content:
          type: string
        contentType:
          type: string
        createdAt:
          type: string
          format: date-time
    TaskDetail:
      type: object
      properties:
        base:
          $ref: '#/components/schemas/TaskBase'
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/TaskComment'
        latestExecutions:
          type: array
          items:
            $ref: '#/components/schemas/AgentExecutionDetail'
        availableActions:
          type: array
          items:
            type: string
        customFields:
          type: object
          additionalProperties: true
    TaskBase:
      type: object
      properties:
        id:
          type: integer
          format: int64
        projectId:
          type: integer
          format: int64
        boardId:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        statusCode:
          type: string
        priority:
          type: string
        aiState:
          type: string
        takeoverState:
          type: string
        assigneeId:
          type: integer
          format: int64
          nullable: true
        version:
          type: integer
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
        assigneeId:
          type: integer
          format: int64
          nullable: true
        dueAt:
          type: string
          format: date-time
          nullable: true
        customFields:
          type: object
          additionalProperties: true
    TransitionTaskRequest:
      type: object
      required: [toStatusCode, version]
      properties:
        toStatusCode:
          type: string
        reason:
          type: string
        version:
          type: integer
    CreateTaskCommentRequest:
      type: object
      required: [content, contentType]
      properties:
        content:
          type: string
        contentType:
          type: string
          enum: [markdown, plain_text]
    TaskTakeoverRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [takeover, release, continue, terminate]
        reason:
          type: string
...
    AgentExecutionLog:
      type: object
      properties:
        logType:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    AgentExecutionDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        taskId:
          type: integer
          format: int64
          nullable: true
        agentId:
          type: integer
          format: int64
        status:
          type: string
        triggerType:
          type: string
        inputPayload:
          type: object
          additionalProperties: true
        outputPayload:
          type: object
          additionalProperties: true
          nullable: true
        errorPayload:
          type: object
          additionalProperties: true
          nullable: true
        costAmount:
          type: number
          format: double
        logs:
          type: array
          items:
            $ref: '#/components/schemas/AgentExecutionLog'
...
security:
  - bearerAuth: []